backend/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3
numpy==1.26.4
xgboost==2.1.1
scikit-learn==1.5.2
optuna==4.1.0
shap==0.46.0
pyarrow==17.0.0
python-multipart==0.0.9
